+++ src/main/java/net/spell_engine/api/item/AttributeResolver.java
@@ -1,10 +1,8 @@
 package net.spell_engine.api.item;
 
-import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.entity.attribute.EntityAttribute;
 import net.minecraft.registry.Registries;
 import net.minecraft.util.Identifier;
-import net.projectile_damage.api.EntityAttributes_ProjectileDamage;
 
 import java.util.HashMap;
 
@@ -20,13 +18,6 @@ public class AttributeResolver {
      * Called upon initialization of this mod.
      */
     private static void setup() {
-        if (FabricLoader.getInstance().isModLoaded("projectile_damage")) {
-            register(
-                    EntityAttributes_ProjectileDamage.attributeId,
-                    EntityAttributes_ProjectileDamage.GENERIC_PROJECTILE_DAMAGE
-            );
-        }
-
         // Mixin here to add custom attributes, don't call `register` outside of this function
         // @Inject(method = "setup", at = @At("TAIL"))
     }